+++ tests/test_dcos_e2e/test_cluster.py
@@ -18,6 +18,7 @@ from py.path import local  # pylint: disable=no-name-in-module, import-error
 
 from dcos_e2e.backends import ClusterBackend
 from dcos_e2e.cluster import Cluster
+from dcos_e2e.node import Output
 
 
 class TestIntegrationTests:
@@ -42,7 +43,7 @@ class TestIntegrationTests:
                 dcos_config=dcos_cluster.base_config,
                 ip_detect_path=cluster_backend.ip_detect_path,
                 build_artifact=oss_artifact,
-                log_output_live=True,
+                output=Output.CAPTURE,
             )
             dcos_cluster.wait_for_dcos_oss()
             yield dcos_cluster
@@ -89,7 +90,7 @@ class TestIntegrationTests:
         pytest_command = ['pytest', '-vvv', '-s', '-x', 'test_auth.py']
         cluster.run_integration_tests(
             pytest_command=pytest_command,
-            log_output_live=True,
+            output=Output.CAPTURE,
         )
 
         # An error is raised with an unsuccessful command.
@@ -97,7 +98,7 @@ class TestIntegrationTests:
             pytest_command = ['pytest', 'test_no_such_file.py']
             result = cluster.run_integration_tests(
                 pytest_command=pytest_command,
-                log_output_live=True,
+                output=Output.CAPTURE,
             )
             # This result will not be printed if the test passes, but it
             # may provide useful debugging information.
@@ -295,8 +296,8 @@ class TestInstallDcosFromPathLogging:
         oss_artifact: Path,
     ) -> None:
         """
-        If `log_output_live` is given as `True`, the installation output is
-        logged live.
+        If ``output`` is given as ``Output.LOG_AND_CAPTURE``, the installation
+        output is logged live.
         """
         with pytest.raises(CalledProcessError):
             # It is not possible to install DC/OS with two master nodes.
@@ -308,7 +309,7 @@ class TestInstallDcosFromPathLogging:
                     build_artifact=oss_artifact,
                     ip_detect_path=cluster_backend.ip_detect_path,
                     dcos_config=cluster.base_config,
-                    log_output_live=True,
+                    output=Output.LOG_AND_CAPTURE,
                 )
 
         assert self._two_masters_error_logged(log_records=caplog.records)