+++ tests/test_dcos_e2e/test_node.py
@@ -5,6 +5,7 @@ See ``test_node_install.py`` for more, related tests.
 """
 
 import logging
+import subprocess
 import sys
 import textwrap
 import uuid
@@ -14,6 +15,7 @@ from subprocess import CalledProcessError, TimeoutExpired
 from typing import Iterator
 
 import pytest
+from _pytest.capture import CaptureFixture
 from _pytest.fixtures import SubRequest
 from _pytest.logging import LogCaptureFixture
 # See https://github.com/PyCQA/pylint/issues/1536 for details on why the errors
@@ -22,7 +24,7 @@ from py.path import local  # pylint: disable=no-name-in-module, import-error
 
 from dcos_e2e.backends import Docker
 from dcos_e2e.cluster import Cluster
-from dcos_e2e.node import Node, Transport
+from dcos_e2e.node import Node, Output, Transport
 
 # We ignore this error because it conflicts with `pytest` standard usage.
 # pylint: disable=redefined-outer-name
@@ -559,7 +561,7 @@ class TestRun:
     # However, we do not skip the whole test so we at least cover more code in
     # the implementation.
     @pytest.mark.parametrize('tty', [True, False])
-    def test_tty(   # pragma: no cover
+    def test_tty(
         self,
         dcos_node: Node,
         tty: bool,
@@ -572,9 +574,9 @@ class TestRun:
             """
             if [ -t 1 ]
             then
-            echo True > {filename}
+            echo True
             else
-            echo False > {filename}
+            echo False
             fi
             """,
         ).format(filename=filename)
@@ -584,16 +586,15 @@ class TestRun:
             shell=True,
         )
 
-        if not sys.stdout.isatty():
+        if not sys.stdout.isatty():  # pragma: no cover
             reason = (
                 'For this test to be valid, stdout must be a TTY. '
                 'Use ``--capture=no / -s`` to run this test.'
             )
             raise pytest.skip(reason)
-
-        assert echo_result.returncode == 0
-        run_result = dcos_node.run(args=['cat', filename])
-        assert run_result.stdout.strip().decode() == str(tty)
+        else:  # pragma: no cover
+            assert echo_result.returncode == 0
+            assert echo_result.stdout.strip().decode() == str(tty)
 
     def test_shell(
         self,
@@ -672,62 +673,145 @@ class TestRun:
         assert echo_result.stdout.strip().decode() == 'hello, world'
         assert echo_result.stderr.strip().decode() == ''
 
-    @pytest.mark.parametrize('shell', [True, False])
-    @pytest.mark.parametrize('log_output_live', [True, False])
-    def test_error(
-        self,
-        caplog: LogCaptureFixture,
-        dcos_node: Node,
-        shell: bool,
-        log_output_live: bool,
-    ) -> None:
+    def test_error(self, dcos_node: Node) -> None:
         """
         Commands which return a non-0 code raise a ``CalledProcessError``.
         """
         with pytest.raises(CalledProcessError) as excinfo:
-            dcos_node.run(
-                args=['rm', 'does_not_exist'],
-                shell=shell,
-                log_output_live=log_output_live,
-            )
+            dcos_node.run(args=['rm', 'does_not_exist'])
 
         exception = excinfo.value
         assert exception.returncode == 1
-        error_message = (
-            'rm: cannot remove ‘does_not_exist’: No such file or directory'
-        )
-        if log_output_live:
-            assert exception.stderr.strip() == b''
-            assert exception.stdout.decode().strip() == error_message
-        else:
-            assert exception.stdout.strip() == b''
-            assert exception.stderr.decode().strip() == error_message
-        # The stderr output is not in the debug log output.
-        debug_messages = set(
-            filter(
-                lambda record: record.levelno == logging.DEBUG,
-                caplog.records,
-            ),
-        )
-        matching_messages = set(
-            filter(
-                lambda record: 'No such file' in record.getMessage(),
-                caplog.records,
-            ),
-        )
-        assert bool(len(debug_messages & matching_messages)) is log_output_live
 
-    def test_log_output_live_and_tty(self, dcos_node: Node) -> None:
-        """
-        A ``ValueError`` is raised if ``tty`` is ``True`` and
-        ``log_output_live`` is ``True``.
-        """
-        with pytest.raises(ValueError) as excinfo:
-            dcos_node.run(
-                args=['echo', '1'],
-                log_output_live=True,
-                tty=True,
-            )
 
-        expected_message = '`log_output_live` and `tty` cannot both be `True`.'
-        assert str(excinfo.value) == expected_message
+class TestOutput:
+    """
+    Tests for the ``output`` parameter of ``Node.run``.
+    """
+
+    def test_default(
+        self,
+        caplog: LogCaptureFixture,
+        dcos_node: Node,
+    ) -> None:
+        """
+        By default, stderr and stdout are captured in the output.
+
+        stderr is logged.
+        """
+        stdout_message = uuid.uuid4().hex
+        stderr_message = uuid.uuid4().hex
+        args = ['echo', stdout_message, '&&', '>&2', 'echo', stderr_message]
+        result = dcos_node.run(args=args, shell=True)
+        assert result.stdout.strip().decode() == stdout_message
+        assert result.stderr.strip().decode() == stderr_message
+
+        args_log, result_log = caplog.records
+
+        assert args_log.levelno == logging.WARNING
+        assert stdout_message in args_log.message
+        assert stderr_message in args_log.message
+        assert 'echo' in args_log.message
+
+        assert result_log.levelno == logging.WARNING
+        assert result_log.message == stderr_message
+
+    def test_capture(
+        self,
+        caplog: LogCaptureFixture,
+        dcos_node: Node,
+    ) -> None:
+        """
+        When given ``Output.CAPTURE``, stderr and stdout are captured in the
+        output.
+
+        stderr is logged.
+        """
+        stdout_message = uuid.uuid4().hex
+        stderr_message = uuid.uuid4().hex
+        args = ['echo', stdout_message, '&&', '>&2', 'echo', stderr_message]
+        result = dcos_node.run(args=args, output=Output.CAPTURE, shell=True)
+        assert result.stdout.strip().decode() == stdout_message
+        assert result.stderr.strip().decode() == stderr_message
+
+        args_log, result_log = caplog.records
+
+        assert args_log.levelno == logging.WARNING
+        assert stdout_message in args_log.message
+        assert stderr_message in args_log.message
+        assert 'echo' in args_log.message
+
+        assert result_log.levelno == logging.WARNING
+        assert result_log.message == stderr_message
+
+    def test_log_and_capture(
+        self,
+        caplog: LogCaptureFixture,
+        dcos_node: Node,
+    ) -> None:
+        """
+        When given ``Output.LOG_AND_CAPTURE``, stderr and stdout are captured
+        in the output as stdout.
+
+        stdout and stderr are logged.
+        """
+        stdout_message = uuid.uuid4().hex
+        stderr_message = uuid.uuid4().hex
+        args = ['echo', stdout_message, '&&', '>&2', 'echo', stderr_message]
+        result = dcos_node.run(
+            args=args,
+            shell=True,
+            output=Output.LOG_AND_CAPTURE,
+        )
+
+        # stderr is merged into stdout.
+        # This is not ideal but for now it is the case.
+        # The order is not necessarily preserved.
+        expected_messages = set([stdout_message, stderr_message])
+        result_stdout = result.stdout.strip().decode()
+        assert set(result_stdout.split('\n')) == expected_messages
+
+        first_log, second_log = caplog.records
+        assert first_log.levelno == logging.DEBUG
+        assert second_log.levelno == logging.DEBUG
+
+        messages = set([first_log.message, second_log.message])
+        assert messages == expected_messages
+
+    def test_no_capture(
+        self,
+        capfd: CaptureFixture,
+        dcos_node: Node,
+    ) -> None:
+        """
+        When given ``Output.NO_CAPTURE``, no output is captured.
+        """
+        stdout_message = uuid.uuid4().hex
+        stderr_message = uuid.uuid4().hex
+        args = ['echo', stdout_message, '&&', '>&2', 'echo', stderr_message]
+        result = dcos_node.run(args=args, shell=True, output=Output.NO_CAPTURE)
+        assert result.stdout is None
+        assert result.stderr is None
+
+        captured = capfd.readouterr()
+        assert captured.out.strip() == stdout_message
+        assert captured.err.strip() == stderr_message
+
+    @pytest.mark.parametrize('output', list(Output))
+    def test_errors(
+        self,
+        caplog: LogCaptureFixture,
+        dcos_node: Node,
+        output: Output,
+    ) -> None:
+        """
+        Errors are always logged at the error level.
+        """
+        args = ['rm', 'does_not_exist']
+        output = Output.CAPTURE
+        with pytest.raises(subprocess.CalledProcessError):
+            dcos_node.run(args=args, shell=True, output=output)
+        [record] = caplog.records
+        assert record.levelno == logging.ERROR
+        expected_message = 'No such file or directory'
+        assert expected_message in record.message