+++ CHANGELOG.rst
@@ -8,6 +8,21 @@ Next
 
 - Add ``dcos-docker create-loopback-sidecar`` and ``dcos-docker destroy-loopback-sidecar`` commands to provide unformatted block devices to DC/OS.
 
+2018.10.17.1
+------------
+
+2018.10.17.0
+------------
+
+- Fix an issue which stopped the SSH transport from working on CLIs.
+
+2018.10.16.0
+------------
+
+- Remove ``log_output_live`` parameters on various functions in favor of new ``output`` options.
+- ``Node.__init__``'s ``ssh_key_path`` parameter now expects a path to an SSH key file with specific permissions.
+   See the documentation for this class for details.
+
 2018.10.13.0
 ------------